+++ tests/test_web.py
@@ -422,6 +422,34 @@ def test_tables_query(
     assert not all(str(row) in all_td for row in df[selected_column][selected_limit:])
 
 
+@pytest.mark.parametrize(
+    ("start_date", "end_date"),
+    [("", ""), ("2025-01-01", ""), ("", "2025-01-31"), ("2025-01-01", "2025-01-31")],
+)
+def test_tables_query_parameters(
+    client: TestClient,
+    sample_config: dict[str, Any],
+    start_date: str,
+    end_date: str,
+) -> None:
+    sql_query = f"select * from {sample_config['tables'][0]['name']} where day between $start_date and $end_date"
+
+    response = client.get(
+        "/tables/query",
+        params={"sql": sql_query, "start_date": start_date, "end_date": end_date},
+    )
+    assert response.status_code == HTTPStatus.OK
+
+    html = response.content.decode()
+    soup = BeautifulSoup(html, "html.parser")
+
+    assert soup.find("h3", string="Parameters")
+    assert soup.find("label", string="start_date")
+    assert soup.find("input", attrs={"name": "start_date", "value": start_date})
+    assert soup.find("label", string="end_date")
+    assert soup.find("input", attrs={"name": "end_date", "value": end_date})
+
+
 def test_tables_query_invalid(
     client: TestClient, delta_table: deltalake.DeltaTable
 ) -> None: